Log method added to occurrence.

+++ copl/com.elster.jupiter.soap.whiteboard.cxf/src/main/java/com/elster/jupiter/soap/whiteboard/cxf/EndPointLog.java
@@ -36,7 +36,7 @@ public interface EndPointLog {
     /**
      * Get id of related occurrence.
      */
-    Integer getOccurrenceId();
+    long getOccurrenceId();
 
     /**
      * Removes a log entry from the system
